SEARCH-8593: Add tests from the HavingStream

+++ solr/solrj/src/java/org/apache/solr/client/solrj/io/ops/EqualsOperation.java
@@ -49,7 +49,7 @@ public class EqualsOperation extends LeafOperation {
       return false;
     }
 
-    return d == val;
+    return d.doubleValue() == val;
   }
 
   public StreamExpression toExpression(StreamFactory factory) throws IOException {
